requirements.txt
mysql-replication
PyMySQL